+++ source/constructs/lib/common-resources/custom-resources/custom-resource-construct.ts
@@ -183,6 +183,7 @@ export class CustomResourcesConstruct extends Construct {
     const bucketDeployment = new BucketDeployment(this, "DeployWebsite", {
       sources: [S3Source.asset(path.join(__dirname, "../../../../demo-ui"))],
       destinationBucket: props.hostingBucket,
+      exclude: ["demo-ui-config.js"],
     });
     Aspects.of(bucketDeployment).add(new ConditionAspect(this.conditions.deployUICondition));
   }